requirement.txt
pyyaml
psycopg2-binary
pgvector
sentence-transformers
requests
torch
transformers
numpy
ollama